+++ src/main/java/survival2d/game/entity/obstacle/Tree.java
@@ -1,14 +1,14 @@
 package survival2d.game.entity.obstacle;
 
-import survival2d.game.entity.base.Circle;
-import survival2d.game.entity.base.HasHp;
 import lombok.Data;
 import org.locationtech.jts.math.Vector2D;
+import survival2d.game.entity.base.Circle;
+import survival2d.game.entity.base.HasHp;
 
 @Data
 public class Tree implements HasHp, Obstacle {
 
-  long id;
+  int id;
   double hp = 100;
   Vector2D position;
   Circle shape;
